ultima

+++ src/pw/SaveBebida.java
@@ -10,17 +10,14 @@ import javax.servlet.http.*;
 
 import com.google.appengine.api.blobstore.BlobKey;
 
-
-
-
 @SuppressWarnings("serial")
 public class SaveBebida extends HttpServlet {
 	public void doPost(HttpServletRequest req, HttpServletResponse resp)
 			throws IOException, ServletException {
 		RequestDispatcher mandar=null;
-		String nombre = req.getParameter("nombreBebida");
+		String 	nombre = req.getParameter("nombreBebida");
 		String tipo = req.getParameter("tipoBebida");
-		int precio = Integer.parseInt(req.getParameter("precioBebida"));
+		Double precio = Double.parseDouble(req.getParameter("precioBebida"));
 		String descripcion = req.getParameter("descripcionBebida");
 
 		Bebida b = new Bebida(nombre, tipo,precio ,descripcion, "null");
